src/vitalDSP/advanced_computation/reinforcement_learning_filter.py
import numpy as np
import random


class ReinforcementLearningFilter:
    """
    A comprehensive Reinforcement Learning Filter for adaptive signal processing.
    This class provides methods to train filters using reinforcement learning algorithms such as
    Q-learning, Deep Q-Networks (DQN), and Proximal Policy Optimization (PPO), and to apply these trained
    filters to signals.

    Methods
    -------
    train_q_learning(episodes=1000, alpha=0.1, gamma=0.99, epsilon=0.1)
        Trains a filter using Q-learning.
    train_dqn(episodes=1000, batch_size=32, gamma=0.99, epsilon=0.1, target_update=10)
        Trains a filter using Deep Q-Networks (DQN).
    train_ppo(epochs=1000, gamma=0.99, lambda_=0.95, clip_ratio=0.2)
        Trains a filter using Proximal Policy Optimization (PPO).
    apply_filter()
        Applies the trained filter to the signal.

    Example Usage
    -------------
    signal = np.sin(np.linspace(0, 10, 100)) + np.random.normal(0, 0.1, 100)
    rl_filter = ReinforcementLearningFilter(signal, action_space=[-1, 0, 1])

    # Train Q-learning based filter
    rl_filter.train_q_learning(episodes=500, alpha=0.1, gamma=0.99, epsilon=0.1)

    # Apply the trained filter
    filtered_signal = rl_filter.apply_filter()
    print("Filtered Signal (Q-Learning):", filtered_signal)

    # Train DQN-based filter
    rl_filter.train_dqn(episodes=500, batch_size=32, gamma=0.99, epsilon=0.1, target_update=10)

    # Apply the trained filter
    filtered_signal_dqn = rl_filter.apply_filter()
    print("Filtered Signal (DQN):", filtered_signal_dqn)

    # Train PPO-based filter
    rl_filter.train_ppo(epochs=500, gamma=0.99, lambda_=0.95, clip_ratio=0.2)

    # Apply the trained filter
    filtered_signal_ppo = rl_filter.apply_filter()
    print("Filtered Signal (PPO):", filtered_signal_ppo)
    """

    def __init__(self, signal, action_space, state_space=None):
        """
        Initialize the ReinforcementLearningFilter class with the signal and action space.

        Parameters
        ----------
        signal : numpy.ndarray
            The signal to be processed.
        action_space : list
            The set of possible actions (e.g., filter adjustments).
        state_space : int or None, optional
            The dimensionality of the state space, if applicable. If None, defaults to the length of the signal.
        """
        self.signal = signal
        self.action_space = action_space
        self.state_space = state_space or len(signal)
        self.q_table = np.zeros(
            (self.state_space, len(action_space))
        )  # Initialize Q-table for Q-learning

    def train_q_learning(self, episodes=1000, alpha=0.1, gamma=0.99, epsilon=0.1):
        """
        Train the filter using Q-learning.

        Q-learning updates a Q-table where each state-action pair has a value. The algorithm iterates over
        episodes and updates the Q-values based on the reward received for each action taken.

        Parameters
        ----------
        episodes : int, optional
            The number of training episodes (default is 1000).
        alpha : float, optional
            The learning rate (default is 0.1).
        gamma : float, optional
            The discount factor (default is 0.99).
        epsilon : float, optional
            The exploration rate for epsilon-greedy policy (default is 0.1).
        """
        for episode in range(episodes):
            state = self._initialize_state()  # Start from the initial state
            for t in range(len(self.signal)):
                # Epsilon-greedy action selection
                if random.uniform(0, 1) < epsilon:
                    action = random.choice(range(len(self.action_space)))  # Explore
                else:
                    action = np.argmax(self.q_table[state])  # Exploit

                # Take action and get next state and reward
                next_state, reward = self._take_action(state, action)
                best_next_action = np.argmax(self.q_table[next_state])

                # Q-learning update rule
                self.q_table[state, action] = self.q_table[state, action] + alpha * (
                    reward
                    + gamma * self.q_table[next_state, best_next_action]
                    - self.q_table[state, action]
                )

                state = next_state  # Move to the next state

    def train_dqn(
        self, episodes=1000, batch_size=32, gamma=0.99, epsilon=0.1, target_update=10
    ):
        """
        Train the filter using Deep Q-Networks (DQN).

        DQN approximates the Q-function using a neural network and uses experience replay to stabilize learning.

        Parameters
        ----------
        episodes : int, optional
            The number of training episodes (default is 1000).
        batch_size : int, optional
            The batch size for experience replay (default is 32).
        gamma : float, optional
            The discount factor (default is 0.99).
        epsilon : float, optional
            The exploration rate for epsilon-greedy policy (default is 0.1).
        target_update : int, optional
            The frequency of updating the target network (default is 10 episodes).
        """
        memory = []
        q_network = self._initialize_neural_network()  # Main Q-network
        target_network = self._initialize_neural_network()  # Target Q-network

        for episode in range(episodes):
            state = self._initialize_state()  # Reset state for each episode
            for t in range(len(self.signal)):
                # Epsilon-greedy action selection
                if random.uniform(0, 1) < epsilon:
                    action = random.choice(range(len(self.action_space)))  # Explore
                else:
                    action = np.argmax(
                        q_network.predict(self._state_to_input(state))
                    )  # Exploit

                # Take action and store the experience
                next_state, reward = self._take_action(state, action)
                memory.append((state, action, reward, next_state))

                # Sample a batch from memory and update the Q-network
                if len(memory) > batch_size:
                    batch = random.sample(memory, batch_size)
                    self._train_neural_network(q_network, batch, gamma, target_network)

                # Update the target network at fixed intervals
                if episode % target_update == 0:
                    target_network.set_weights(q_network.get_weights())

                state = next_state  # Move to the next state

    def train_ppo(self, epochs=1000, gamma=0.99, lambda_=0.95, clip_ratio=0.2):
        """
        Train the filter using Proximal Policy Optimization (PPO).

        PPO is a policy gradient method that seeks to optimize policies while ensuring
        the updates don't deviate too much from the previous policy.

        Parameters
        ----------
        epochs : int, optional
            The number of training epochs (default is 1000).
        gamma : float, optional
            The discount factor (default is 0.99).
        lambda_ : float, optional
            The GAE-lambda parameter (default is 0.95).
        clip_ratio : float, optional
            The clipping parameter for PPO (default is 0.2).
        """
        policy_network = self._initialize_policy_network()
        value_network = self._initialize_value_network()

        for epoch in range(epochs):
            # Collect trajectories for policy updates
            trajectories = self._collect_trajectories(
                policy_network, value_network, gamma, lambda_
            )
            self._update_policy_network(policy_network, trajectories, clip_ratio)
            self._update_value_network(value_network, trajectories)

    def apply_filter(self):
        """
        Apply the trained filter to the signal using the learned policy or Q-values.

        Returns
        -------
        numpy.ndarray
            The filtered signal.
        """
        filtered_signal = np.zeros_like(self.signal)
        state = self._initialize_state()

        for t in range(len(self.signal)):
            action = np.argmax(
                self.q_table[state]
            )  # Select the best action based on Q-values
            filtered_signal[t] = self._apply_action(action, t)
            state = self._update_state(state, action)

        return filtered_signal

    # Helper Methods for Q-Learning

    def _initialize_state(self):
        """Initialize the state for Q-learning."""
        return 0

    def _take_action(self, state, action):
        """Take an action and return the next state and reward."""
        next_state = (state + 1) % self.state_space
        reward = self._calculate_reward(state, action)
        return next_state, reward

    def _calculate_reward(self, state, action):
        """Calculate the reward based on the current state and action."""
        return -abs(self.signal[state] - self._apply_action(action, state))

    def _apply_action(self, action, t):
        """Apply an action to the signal at time t."""
        return self.signal[t] + self.action_space[action]

    def _update_state(self, state, action):
        """Update the state based on the action taken."""
        return (state + 1) % self.state_space

    # Helper Methods for DQN

    def _initialize_neural_network(self):
        """Initialize a simple neural network for DQN."""
        return SimpleNeuralNetwork(self.state_space, len(self.action_space))

    def _state_to_input(self, state):
        """Convert the state index to an input vector for the neural network."""
        state_input = np.zeros(self.state_space)
        state_input[state] = 1
        return state_input

    def _train_neural_network(self, network, batch, gamma, target_network):
        """Train the neural network using a batch of experiences."""
        for state, action, reward, next_state in batch:
            target = reward + gamma * np.max(
                target_network.predict(self._state_to_input(next_state))
            )
            network.train(self._state_to_input(state), action, target)

    # Helper Methods for PPO

    def _initialize_policy_network(self):
        """Initialize the policy network for PPO."""
        return SimplePolicyNetwork(self.state_space, len(self.action_space))

    def _initialize_value_network(self):
        """Initialize the value network for PPO."""
        return SimpleValueNetwork(self.state_space)

    def _collect_trajectories(self, policy_network, value_network, gamma, lambda_):
        """Collect trajectories for PPO training."""
        trajectories = []
        state = self._initialize_state()
        for t in range(len(self.signal)):
            state_input = self._state_to_input(state)
            action = policy_network.sample_action(state_input)
            next_state, reward = self._take_action(state, action)
            advantage = self._calculate_advantage(
                state, action, reward, next_state, value_network, gamma, lambda_
            )
            trajectories.append((state, action, reward, advantage))
            state = next_state
        return trajectories

    def _calculate_advantage(
        self, state, action, reward, next_state, value_network, gamma, lambda_
    ):
        """Calculate the advantage estimate for PPO."""
        value = value_network.predict(self._state_to_input(state))
        next_value = value_network.predict(self._state_to_input(next_state))
        td_error = reward + gamma * next_value - value
        advantage = td_error + lambda_ * (reward + gamma * next_value - value)
        return advantage

    def _update_policy_network(self, policy_network, trajectories, clip_ratio):
        """Update the policy network using PPO."""
        for state, action, _, advantage in trajectories:
            state_input = self._state_to_input(state)
            old_prob = policy_network.predict(state_input)[action]
            new_prob = policy_network.train(state_input, action, advantage)
            ratio = new_prob / (
                old_prob + 1e-8
            )  # Added epsilon to prevent division by zero
            clipped_ratio = np.clip(ratio, 1 - clip_ratio, 1 + clip_ratio)
            objective = np.minimum(ratio * advantage, clipped_ratio * advantage)
            policy_network.update(state_input, action, objective)

    def _update_value_network(self, value_network, trajectories):
        """Update the value network using PPO."""
        for state, _, reward, _ in trajectories:
            value_network.train(self._state_to_input(state), reward)


# Simple Neural Network Class for DQN (Placeholder)


class SimpleNeuralNetwork:
    def __init__(self, input_size, output_size):
        self.weights = np.random.randn(input_size, output_size)

    def predict(self, state):
        return np.dot(state, self.weights)

    def train(self, state, action, target):
        self.weights[:, action] += 0.01 * (target - self.predict(state)[action]) * state

    def get_weights(self):
        return self.weights

    def set_weights(self, weights):
        self.weights = weights


# Simple Policy Network Class for PPO (Placeholder)


class SimplePolicyNetwork:
    def __init__(self, input_size, output_size):
        self.weights = np.random.randn(input_size, output_size)

    def predict(self, state):
        logits = np.dot(state, self.weights)
        probabilities = np.exp(logits) / np.sum(np.exp(logits))
        probabilities = np.clip(probabilities, 1e-5, 1 - 1e-5)
        return probabilities

    def sample_action(self, state):
        probabilities = self.predict(state)
        return np.argmax(np.random.multinomial(1, probabilities))

    def train(self, state, action, advantage):
        self.weights[:, action] += 0.01 * advantage * state
        return self.predict(state)[action]

    def update(self, state, action, objective):
        self.weights[:, action] += 0.01 * objective * state


# Simple Value Network Class for PPO (Placeholder)
class SimpleValueNetwork:
    def __init__(self, input_size):
        """
        Initialize the value network with random weights.

        Parameters
        ----------
        input_size : int
            The size of the input layer (state space dimension).
        """
        self.weights = np.random.randn(input_size)

    def predict(self, state):
        """
        Predict the value of a given state.

        Parameters
        ----------
        state : numpy.ndarray
            The input state vector.

        Returns
        -------
        float
            The predicted value of the state.
        """
        return np.dot(state, self.weights)

    def train(self, state, reward):
        """
        Train the value network by updating weights based on the reward.

        Parameters
        ----------
        state : numpy.ndarray
            The input state vector.
        reward : float
            The reward value used for training.
        """
        prediction = self.predict(state)
        self.weights += 0.01 * (reward - prediction) * state
